% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.dynamicLM}
\alias{coef.dynamicLM}
\title{Get the coefficients of a fitted supermodel in dynamicLM}
\usage{
\method{coef}{dynamicLM}(object, ...)
}
\arguments{
\item{object}{}
}
\value{
Vector of coefficients for a Cox landmark supermodel or list of
coefficients for each cause-specific model for a CSC landmark supermodel.
}
\description{
Get the coefficients of a fitted supermodel in dynamicLM
}
