% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CSC.fixed.coefs}
\alias{CSC.fixed.coefs}
\title{Altered code from \code{riskRegression} of the cause-specific Cox model
to fit a CSC model with given coefficients.}
\usage{
CSC.fixed.coefs(formula, data, cause, cause.specific.coefs, ...)
}
\arguments{
\item{formula}{Formula to fit the model}

\item{data}{Data on which to which}

\item{cause}{Main cause of interest}

\item{cause.specific.coefs}{Coefficients that each model should be fit with}

\item{...}{Additional arguments to coxph.}
}
\value{
CSC model
}
\description{
Altered code from \code{riskRegression} of the cause-specific Cox model
to fit a CSC model with given coefficients.
}
\references{
\itemize{
\item \code{riskRegression} package:
\url{https://cran.r-project.org/web/packages/riskRegression/index.html}
}
}
