% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.penLM.R
\name{cv.penLM}
\alias{cv.penLM}
\title{Cross-validation for a  regular or cause-specific Cox landmark supermodel
with lasso or elasticnet penalization.}
\usage{
cv.penLM(
  x,
  y,
  lmdata,
  xcols,
  id_col,
  alpha = 1,
  nfolds = 10,
  type.measure = "deviance",
  seed = NULL,
  foldid = NULL,
  ...
)
}
\arguments{
\item{x}{Input matrix with each row an observation.}

\item{y}{Response variable: either a \code{Surv} or \code{Hist} object.}

\item{lmdata}{An object of class "LMdataframe", this can be created by
running \code{\link[=stack_data]{stack_data()}} and \code{\link[=add_interactions]{add_interactions()}}}

\item{xcols}{A vector of column names of the data stored in \code{lmdata} that
are to be used as dependent variables. If not specified, it is assumed that
all non-response variables are the dependent variables.}

\item{id_col}{Column name or index that identifies individuals in data.
Used to ensure the same individuals appear in the different cross
validation sets.}

\item{alpha}{The elastic net mixing parameter: Lies betweent 0 and 1. At 1,
the penalty is the LASSO penalty, and at 0, the penalty is the ridge
penalty.}

\item{nfolds}{Number of folds in k-fold cross validation. Default is 10.}

\item{type.measure}{Loss for cross-validation. Currently the only option is
"deviance" which is the partial-likelihood for the Cox model. If using
cause-specific Cox models, this is evaluated on each model separately.}

\item{seed}{Set a seed.}

\item{foldid}{TODO}

\item{...}{Additional arguments to \code{cv.glmnet}.}
}
\value{
An object of class \code{cv.penLM}. This is a list of \code{cv.glmnet} objects
(one for each cause-specific Cox model or a list of length one for a
regular Cox model). The object also has attributes \code{survival.type}
(\code{competing.risk} or \code{survival}) and \code{lmdata} and \code{xcols} which store the
inputs if given.
Functions \code{print} and \code{plot} exist for the object. To make predictions,
see \link{dynamic_lm.cv.penLM}.
}
\description{
Fit by calling \code{\link[=cv.glmnet]{cv.glmnet()}}. As in \code{cv.glmnet}, k-fold cross validation is
performed. This produces a plot and returns a value for lambda. Input can be
as typically done for \code{cv.glmnet} in the form of \code{x} and \code{y} or with a
landmark super dataset \code{lmdata} specifying dependent columns in \code{xcols}.
}
\examples{
\dontrun{
}
}
