% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_lm.R
\name{dynamic_lm.data.frame}
\alias{dynamic_lm.data.frame}
\title{Fit a dynamic Cox or cause-specific Cox landmark supermodel to a dataframe.}
\usage{
\method{dynamic_lm}{data.frame}(
  lmdata,
  formula,
  type = "coxph",
  method = "breslow",
  func_covars,
  func_lms,
  lm_col,
  outcome,
  w,
  lm_covs,
  cluster,
  x = FALSE,
  ...
)
}
\arguments{
\item{lmdata}{A dataframe that should be a stacked dataset across landmark
times.}

\item{formula}{The formula to be used, remember to include "+cluster(ID)" for
the column that indicates the ID of the individual for robust error
estimates.}

\item{type}{"coxph" or "CSC"/"CauseSpecificCox"}

\item{method}{A character string specifying the method for tie handling.
Default is "breslow". More information can be found in \code{\link[survival:coxph]{survival::coxph()}}.}

\item{func_covars}{A list of functions to use for interactions between LMs
and covariates.}

\item{func_lms}{A list of functions to use for transformations of the
landmark times.}

\item{lm_col}{Character string specifying the column name that indicates the
landmark time point for a row.}

\item{outcome}{List with items time and status, containing character strings
identifying the names of time and status variables, respectively, of the
survival outcome}

\item{w}{Scalar, the value of the prediction window (ie predict w-year/other
time period risk from the LM points)}

\item{lm_covs}{Vector of strings indicating the columns that are to have a
LM interaction}

\item{cluster}{Variable which clusters the observations (for e.g., identifies
repeated patient IDs), for the purposes of a robust variance. If omitted,
extracted from \code{formula}.}

\item{x}{Logical value. If set to true, \code{lmdata} is stored in the returned
object. This is required for internal validation.}

\item{...}{Arguments given to coxph or CSC.}
}
\value{
An object of class "LMcoxph" or "LMCSC" with components:
\itemize{
\item model: fitted model
\item type: as input
\item w, func_covars, func_lms, lm_covs, all_covs, outcome: as in input.
\item LHS: the survival outcome
\item linear.predictors: the vector of linear predictors, one per subject.
Note that this vector has not been centered.
\item args: arguments used to call model fitting
\item id_col: the cluster argument, often specifies the column with patient ID
\item lm_col: column name that indicates the landmark time point for a row.
}
}
\description{
Note that it is recommended to rather use \code{\link[=stack_data]{stack_data()}} and
\code{\link[=add_interactions]{add_interactions()}} to create an object of class LMdataframe rather than
directly calling \code{dynamic_lm} on a dataframe to ensure the data has the
correct form.
}
\examples{
\dontrun{
}
}
