% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_lm.R
\name{dynamic_lm.formula}
\alias{dynamic_lm.formula}
\title{Fit a dynamic Cox or cause-specific Cox landmark supermodel}
\usage{
\method{dynamic_lm}{formula}(formula, lmdata, type, ...)
}
\arguments{
\item{formula}{The formula to be used, remember to include \code{+cluster(ID)} for
the column that indicates the ID of the individual for robust error
estimates. See details for further information.
Note that transformations (e.g., \code{x1*x2}) cannot be used in the formula and
factors/categorical variables must first be made into dummy variables.}

\item{lmdata}{An object of class "LMdataframe", this can be created by
running \code{\link[=stack_data]{stack_data()}} and \code{\link[=add_interactions]{add_interactions()}}}

\item{type}{"coxph" or "CSC"/"CauseSpecificCox"}

\item{...}{Arguments given to coxph or CSC.}
}
\value{
An object of class "LMcoxph" or "LMCSC" with components:
\itemize{
\item model: fitted model
\item type: as input
\item w, func_covars, func_lms, lm_covs, all_covs, outcome: as in \code{lmdata}
\item LHS: the survival outcome
\item linear.predictors: the vector of linear predictors, one per subject.
Note that this vector has not been centered.
\item args: arguments used to call model fitting
\item id_col: the cluster argument, often specifies the column with patient ID
\item lm_col: column name that indicates the landmark time point for a row.
}
}
\description{
Fit a dynamic Cox or cause-specific Cox landmark supermodel
}
\details{
For standard survival data (one event and possible censoring), use
\code{type = "coxph"} and a a formula with left-hand side (LHS) of the form
\code{Surv(LM, Time, event)}. For competing risks (multiple events and possible
censoring), use \code{type = "CSC"} and a LHS of the form
\code{Hist(Time, event, LM)}. This form is kept to ensure compatibility with the
original dynamicLM library, although in later versions, the formula is the
second argument.
}
