% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLM.R
\name{fitLM.cv.penLM}
\alias{fitLM.cv.penLM}
\title{Fit a penalized cross-validated coxph or CSC super model}
\usage{
\method{fitLM}{cv.penLM}(object, lambda = "lambda.1se", ...)
}
\arguments{
\item{object}{A fitted object of class "cv.LMpen". This can be created by
calling \code{cv.penLM} using arguments \code{LMdata} and \code{xcols}.}

\item{lambda}{Value of the penalty parameter \code{lambda} at which to fit a model.
Default is "lambda.1se" stored in the object; "lambda.min" can also be used
or a specific value of can be input. For cause-specific Cox super models,
this must be a list or vector of values: one for each cause.}
}
\value{
An object of class "penLMcoxph" or "penLMCSC" with components:
\itemize{
\item model: fitted model
\item type: as input
\item w, func_covars, func_LMs, LMcovars, allLMcovars, outcome: as in LMdata
\item LHS: the LHS of the input formula
\item linear.predictors: the vector of linear predictors, one per subject. Note
that this vector has not been centered.
\item s: the values of lambda for which this model has been fit.
}
}
\description{
Fit a penalized cross-validated coxph or CSC super model
}
\details{
The Breslow method is used for handling ties, as we use the \code{glmnet}
package which does the same.
}
\examples{
\dontrun{
}
}
