% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLM.R
\name{fitLM.LMpen}
\alias{fitLM.LMpen}
\title{Fit a penalized coxph or CSC model for a specific coefficient}
\usage{
<<<<<<< HEAD
\method{fitLM}{LMpen}(object, id, s, ...)
=======
\method{fitLM}{LMpen}(object, id, s)
>>>>>>> 56371f8 (Merge conflicts)
}
\arguments{
\item{object}{A fitted object of class "LMpen" (i.e., TODO), this can be
created by calling \code{penLM}. This should be fit using arguments \code{LMdata} and
\code{xcols}.}

\item{id}{Column name that indicates the ID of the individual for robust
error estimates.}

\item{s}{Value of the penalty parameter \code{lambda} at which to fit a model.}
}
\value{
An object of class "penLMcoxph" or "penLMCSC" with components:
\itemize{
\item model: fitted model
\item type: as input
\item w, func_covars, func_LMs, LMcovars, allLMcovars, outcome: as in LMdata
\item LHS: the LHS of the input formula
\item linear.predictors: the vector of linear predictors, one per subject. Note
that this vector has not been centered.
\item TODO: add regularization information
}
}
\description{
Fit a penalized coxph or CSC model for a specific coefficient
}
\details{
The Breslow method is used for handling ties, as we use the \code{glmnet}
package which does the same.
}
\examples{
\dontrun{
}
}
