% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLM.R
\name{fitLM}
\alias{fitLM}
\title{Fit a landmark supermodel}
\usage{
fitLM(...)
}
\value{
An object of class "LMcoxph", "LMCSC", "penLMcoxph" or "penLMCSC"
with components:
\itemize{
\item model: fitted model
\item type: as input
\item w, func_covars, func_LMs, LMcovars, allLMcovars, outcome: as in LMdata
\item LHS: the LHS of the input formula
<<<<<<< HEAD
\item linear.predictors: the vector of linear predictors, one per subject. Note
that this vector has not been centered.
=======
\item linear.predictors: the vector of linear predictors, one per subject. Note that this vector has not been centered.
\item args: arguments used to call model fitting
>>>>>>> f28317e (LMcalPlot bootstrapping v1 complete + minor updates)
}
}
\description{
Fit a Cox or cause-specific Cox landmark supermodel, with or without
regularization.
}
\details{
To fit coxph or cause-specific Cox models on a landmark super
dataset, see \code{fitLM.LMdataframe}. To fit penalized Cox or penalized
cause-specific Cox models see \code{fitLM.LMpen}.
}
\seealso{
\code{fitLM.LMdataframe}, \code{fitLM.LMpen}
}
