% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pbc_long.R
\name{get_pbc_long}
\alias{get_pbc_long}
\title{Get the PBC data in long data from the survival package}
\usage{
get_pbc_long()
}
\value{
A data frame with the PBC data in long format
}
\description{
This function prepares the PBC data from the \code{survival} package in long
format, merging baseline data with longitudinal data and cleaning it for
analysis.
}
\examples{
pbc_df <- get_pbc_long()
head(pbc_df)

}
