% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMcalPlot.R
\name{LMcalPlot}
\alias{LMcalPlot}
\title{Calibration plots for dynamic risk prediction landmark models}
\usage{
LMcalPlot(
  object,
  times,
  formula,
  data,
  tLM,
  ID_col = "ID",
  split.method = "none",
  B = 1,
  M,
  unit = "year",
  cause,
  plot = T,
  main,
  sub = T,
  ...
)
}
\arguments{
\item{object}{A named list of prediction models, where allowed entries are outputs from predLMrisk or supermodels from fitLM}

\item{times}{Landmark times for which calibration must be plot. These must be a subset of LM times used during the prediction}

\item{formula}{A survival or event history formula (Hist(..)). The left hand side is used to compute the expected event status.
If none is given, it is obtained from the prediction object.}

\item{data}{Data on which to predict if the object is a fitted model. If bootstrapping, must be an object
of class LMdataframe. If not bootstrapping, it can be a dataframe.}

\item{tLM}{Landmark times corresponding to the patient entries in data. Only required if data is a dataframe.
tLM can be a string (indicating a column in data), a vector of length nrow(data),
or a single value if all patient entries were obtained at the same landmark time.}

\item{ID_col}{Column name that identifies individuals in data. Only required if bootstrapping.}

\item{split.method}{Defines the internal validation design as in pec::calPlot. Options are currently none or bootcv}

\item{B}{Number of times bootstrapping is performed.}

\item{M}{Subsample size for training in cross-validation. Entries not sampled in the M subsamples are used for validation.}

\item{unit}{The unit of w, i.e. w-unit prediction ("year","month", etc...). Only used to label the plot.}

\item{cause}{Cause of interest if considering competing risks}

\item{plot}{If FALSE, do not plot the results, just return a plottable object. Default is TRUE.}

\item{main}{Optional title to override default.}

\item{sub}{If TRUE, add a subheading with the number of individuals at risk, and the number that under the event of interest.
Default is TRUE. Set to FALSE for bootstrapping.}

\item{...}{Additional arguments to pass to calPlot}
}
\value{
List of plots of w-year risk, one entry per prediction/landmark time point
}
\description{
Calibration plots for dynamic risk prediction landmark models
}
\details{
Most errors in plotting occur when a formula is not given. Formulas can look like \code{Surv(LM,Time,event)~1} / \code{Surv(LM,Time,event==1)~1} / \code{Hist(Time,event,LM)~1} / similar...

See the Github for example code on using LMcalPlot in general.
}
\examples{
\dontrun{
par(mfrow=c(2,2),pty="s")
outlist = LMcalPlot(list("Model1"=p1),
                    unit="month",            # for the title
                    times=c(6,12,18,24),       # landmarks at which to provide calibration plots
                    formula="Hist(event,Time,LM)~1",
                    method="quantile", q=10, # method for calibration plot
                    ylim=c(0,0.4), xlim=c(0,0.4))
}
}
