% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMScore.R
\name{LMScore}
\alias{LMScore}
\title{Methods (AUCt, Brier) to score the predictive performance of dynamic risk markers from LM super models}
\usage{
LMScore(
  preds,
  formula,
  metrics = c("auc", "brier"),
  cause,
  tLM,
  unit,
  split.method,
  B,
  M,
  ...
)
}
\arguments{
\item{preds}{A named list of prediction models, where allowed entries are outputs from predLMrisk}

\item{formula}{A survival or event history formula. The left hand side is used to compute the expected event status.
If none is given, it is obtained from the prediction object.}

\item{metrics}{Character vector specifying which metrics to apply. Choices are "auc" and "brier". Case matters.}

\item{cause}{Cause of interest if considering competing risks. If left blank, this is inferred from the preds object.}

\item{tLM}{Landmark times for which scores must be given. These must be a subset of LM times used during the prediction.
If left blank, all landmark prediction time points are considered.}

\item{unit}{Time unit for window of prediction, e.g., "year", "month", etc. Only used for printing results.}

\item{split.method}{Defines the internal validation design as in riskRegression::Score. Options are currently none or bootcv}

\item{B}{The number of bootstrap steps for cross-validation.}

\item{M}{The size of the subsamples drawn for bootstrap cross-validation.
If specified it has to be an integer smaller than the size of data.}

\item{...}{Additional arguments to pass to Score (riskRegression package).
These arguments have been included for user flexibility but have not been tested and are not necessarily appropriate.}
}
\value{
An object of class "LMScore", which has components:
\itemize{
\item auct: dataframe containing time-dependent auc information if "auc" was a metric
\item briert: dataframe containing time-dependent brier score if "brier" was a metric
}
}
\description{
Methods (AUCt, Brier) to score the predictive performance of dynamic risk markers from LM super models
}
\details{
See the Github for example code
}
