% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penLM.R
\name{penLM}
\alias{penLM}
\title{Fit a regular or cause-specific Cox landmark supermodel with lasso or
elastic-net penalization.}
\usage{
penLM(x, y, lmdata, xcols, alpha = 1, ...)
}
\arguments{
\item{x}{Input matrix with each row an observation.}

\item{y}{Response variable: either a \code{Surv} or \code{Hist} object.}

\item{lmdata}{An object of class "LMdataframe", this can be created by
running \code{\link[=stack_data]{stack_data()}} and \code{\link[=add_interactions]{add_interactions()}}}

\item{xcols}{A vector of column names of the data stored in \code{lmdata} that
are to be used as dependent variables. If not specified, it is assumed that
all non-response variables are the dependent variables.}

\item{alpha}{The elastic net mixing parameter: Lies betweent 0 and 1. At 1,
the penalty is the LASSO penalty, and at 0, the penalty is the ridge
penalty.}

\item{...}{Additional arguments passed to \code{\link[=glmnet]{glmnet()}}.}
}
\value{
An object of class \code{penLM}. This is a list of \code{glmnet} objects (one
for each cause-specific Cox model or a list of length one for a regular Cox
model). The object also has attributes \code{survival.type} (\code{competing.risk}
or \code{survival}) and \code{lmdata} and \code{xcols} which store the inputs if given.
Functions \code{print} and \code{plot} exist for the object. To make predictions,
see \code{\link[=dynamic_lm.penLM]{dynamic_lm.penLM()}} and \code{\link[=predict]{predict()}}.
}
\description{
Fit by calling \verb{[glmnet()]}. As in \code{glmnet}, the model is fit via penalized
maximum likelihood to produce a regularization path at a grid of values for
the regularization parameter lambda. Input can be as typically done for
\code{glmnet} in the form of \code{x} and \code{y} or with a landmark super dataset \code{lmdata}
specifying dependent columns in \code{xcols}.
}
\examples{
\dontrun{
}
}
