% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.coefs}
\alias{plot.coefs}
\title{Generic function to plot coefficients}
\usage{
\method{plot}{coefs}(
  x,
  single_plot = TRUE,
  max_coefs = NULL,
  col = "blue",
  xlab = "Coefficient value",
  ...
)
}
\arguments{
\item{x}{(Named) Vector of coefficients}

\item{single_plot}{Logical, defaults to TRUE. A single plot for both
positive and negative coefficients, or two separate plots.}

\item{max_coefs}{Default is to plot all coefficients. If specified, gives the
maximum number of coefficients to plot.}

\item{col}{Fill color for the barplot.}

\item{xlab}{x-axis Label}

\item{...}{Additional arguments to barplot.}
}
\description{
Can plot positive and negative coefficients in two separate plots or the
same. X-axes are the same if separate plots are used.
}
