% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cv.penLM}
\alias{plot.cv.penLM}
\title{Plot cross-validation curve created by \code{\link[=cv.penLM]{cv.penLM()}}, analogous to plotting
from \code{\link[=cv.glmnet]{cv.glmnet()}}}
\usage{
\method{plot}{cv.penLM}(
  x,
  all_causes = FALSE,
  silent = FALSE,
  label = FALSE,
  sign.lambda = 1,
  se.bands = TRUE,
  ...
)
}
\arguments{
\item{x}{a fitted \code{\link[=cv.penLM]{cv.penLM()}} object}

\item{all_causes}{if penLM fit a cause-specific Cox model, set TRUE to plot
coefficient profile plots for each model.}

\item{silent}{Set TRUE to hide messages.}

\item{label}{Set TRUE to label the curves by variable index numbers.}

\item{sign.lambda}{Plot against \code{log(lambda)} (default) or its negative
if set to -1.}

\item{se.bands}{Logical. If TRUE, shading is produced to show stand-error
bands. Defaults to TRUE.}

\item{\dots}{additional graphical parameters}
}
\description{
The cross-validation curve is plotted as a function of the lambda values
used. Upper and lower standard deviation is plotted too.
}
\details{
If the model is a survival model (i.e., no competing risks), then the output
is the same as a call to \code{cv.glmnet} would produce. For competing risks, the
default is only to plot the cross-validation curve for the cause of
interest (first cause) Further events can be examined by setting
\code{all_causes = TRUE}.
}
