% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.dynamicLM}
\alias{plot.dynamicLM}
\title{Plots the dynamic log-hazard ratio of a cox or CSC supermodel}
\usage{
\method{plot}{dynamicLM}(
  x,
  covars,
  conf_int = TRUE,
  cause,
  end_time,
  logHR = TRUE,
  extend = FALSE,
  silence = FALSE,
  xlab = "Landmark time",
  ylab,
  ylim,
  main,
  ...
)
}
\arguments{
\item{x}{A fitted supermodel}

\item{covars}{Vector or list of strings indicating the variables to plot
(note these must be given without time interaction).}

\item{conf_int}{Include confidence intervals or not, default is TRUE}

\item{cause}{Cause of interest if considering competing risks}

\item{end_time}{Final time point to plot HR, defaults to the last landmark
point used in model fitting.}

\item{logHR}{Boolean, if true plots the log of the hazard ratio, if false
plots the hazard ratio. Default is TRUE.}

\item{extend}{Argument to allow for HR to be plot at landmark times that are
later than the LMs used in model fitting.
Default is FALSE. If set to TRUE, the HR may be unreliable.}

\item{silence}{silence the warning message when end_time > LMs used in
fitting the model}

\item{xlab}{x label for the plots}

\item{ylab}{y label for the plots}

\item{ylim}{y limit for the plots}

\item{main}{Vector of strings indicating the title of each plot. Must be in
the same order as covars.}

\item{...}{Additional arguments passed to plot}

\item{discrete_grid}{Defaults to 0.1, how to discretize the grid for plotting}
}
\value{
Plots for each variable in covars showing the dynamic hazard ratio
}
\description{
Plots the dynamic log-hazard ratio of a cox or CSC supermodel
}
\details{
See our \href{https://github.com/thehanlab/dynamicLM}{GitHub} for example
code
}
\examples{

data(relapse)
outcome <- list(time = "Time", status = "event")
covars <- list(fixed = c("male", "stage", "bmi"),
               varying = c("treatment"))
w <- 60; lms <- c(0, 6, 12, 18)
lmdata <- stack_data(relapse, outcome, lms, w, covars, format = "long",
                     id = "ID", rtime = "T_txgiven")
lmdata <- add_interactions(lmdata, func_covars = c("linear", "quadratic"),
                           func_lms = c("linear", "quadratic"))
formula <- "Hist(Time, event, LM) ~ male + male_LM1 + male_LM2 +
            stage + stage_LM1 + stage_LM2 + bmi + bmi_LM1 + bmi_LM2 +
            treatment + treatment_LM1 + treatment_LM2 + LM1 + LM2 + cluster(ID)"
supermodel <- dynamic_lm(lmdata, as.formula(formula), "CSC", x = TRUE)

par(mfrow = c(2, 3))
plot(supermodel)

par(mfrow = c(1, 2))
plot(supermodel,
     covars = c("stage", "bmi"), # subset of covariates to plot
     logHR = FALSE,              # plot HR instead of log HR
     conf_int = FALSE,           # do not plot confidence intervals
     main = c("HR of stage", "HR of BMI"))

}
