% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.LMcalibrationPlot}
\alias{plot.LMcalibrationPlot}
\title{Plot an object output from \code{\link[=calplot]{calplot()}}: plot the calibration plots.}
\usage{
\method{plot}{LMcalibrationPlot}(object)
}
\arguments{
\item{object}{An object of class "LMcalibrationPlot" output from \code{\link[=calplot]{calplot()}}}

\item{...}{}
}
\description{
Plot an object output from \code{\link[=calplot]{calplot()}}: plot the calibration plots.
}
