% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.LMpen}
\alias{plot.LMpen}
\title{Plot coefficients from an object created by calling \code{penLM}}
\usage{
\method{plot}{LMpen}(x, xvar = "norm", all_causes = FALSE, silent = FALSE, label = FALSE, ...)
}
\arguments{
\item{x}{a fitted penLM object}

\item{xvar}{As in glmnet: "What is on the X-axis. "\code{norm}" plots against the
L1-norm of the coefficients, "\code{lambda}" against the log-lambda sequence,
and "\code{dev}" against the percent deviance explained."}

\item{all_causes}{if penLM fit a cause-specific Cox model, set TRUE to plot
coefficient profile plots for each model.}

\item{silent}{Set TRUE to hide messages.}

\item{label}{Set TRUE to label the curves by variable index numbers.}

\item{\dots}{additional graphical parameters}
}
\description{
As in the glmnet package, produces a coefficient profile plot of the
coefficient paths.
}
\details{
If the model is a survival model (i.e., no competing risks), then the output
is the same as a call to glmnet would produce. For competing risks, the
default is only to plot the coefficient profile plot for the cause of
interest (first cause) Further events can be examined by setting
\code{all_causes = TRUE}.
}
