% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.LMScore}
\alias{plot.LMScore}
\title{Plot an object output from \code{\link[=score]{score()}}: plot the landmark and time-dependent
Brier and/or AUC of dynamic landmark supermodels.}
\usage{
\method{plot}{LMScore}(
  x,
  metrics,
  contrasts = FALSE,
  landmarks = TRUE,
  summary = FALSE,
  se = TRUE,
  pairwise_contrasts = FALSE,
  cutoff_contrasts = 0.05,
  pairwise_heights,
  width,
  loc,
  xlab,
  ylab,
  pch,
  ylim,
  xlim,
  main,
  col = NULL,
  cex = 1,
  length = 0.1,
  legend = TRUE,
  legend.title = NULL,
  auc = TRUE,
  brier = TRUE,
  ...
)
}
\arguments{
\item{x}{An object of class "LMScore" output from \code{\link[=score]{score()}}}

\item{metrics}{One or both of "AUC" and "Brier"}

\item{contrasts}{Plot the difference between metrics. Default is
FALSE and plots the metrics themselves.}

\item{landmarks}{TODO}

\item{summary}{Plot the summary metric. Default is FAULT and plots
landmark-specific metrics.}

\item{se}{Boolean, default TRUE. To include point wise confidence intervals.}

\item{pairwise_contrasts}{TODO}

\item{cutoff_contrasts}{TODO}

\item{pairwise_heights}{TODO}

\item{width}{TODO}

\item{loc}{Location for legend.}

\item{xlab, ylab, pch, ylim, xlim, main}{graphical parameters}

\item{col}{TODO}

\item{cex}{TODO (for legend)}

\item{length}{The width of the ends of the error bars.}

\item{legend}{Include a legend or not. Default is TRUE.}

\item{auc}{Plot the AUC or not (if available). Default is TRUE.}

\item{brier}{Plot the Brier Score or not (if available). Default is TRUE.}

\item{...}{Additional arguments to \code{plot()}}
}
\description{
Plot an object output from \code{\link[=score]{score()}}: plot the landmark and time-dependent
Brier and/or AUC of dynamic landmark supermodels.
}
