% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.penLMcoxph}
\alias{plot.penLMcoxph}
\title{Plot the non-zero coefficients of a penalized Cox landmark supermodel}
\usage{
\method{plot}{penLMcoxph}(
  x,
  single_plot = TRUE,
  max_coefs = NULL,
  col = "blue",
  xlab = "Coefficient value",
  ...
)
}
\arguments{
\item{x}{a penalized Cox supermodel - created by calling \code{fitLM} on an
object created from \code{penLM}/\code{cv.pen}}

\item{single_plot}{Logical, defaults to TRUE. A single plot for both
positive and negative coefficients, or two separate plots.}

\item{max_coefs}{Default is to plot all coefficients. If specified, gives the
maximum number of coefficients to plot.}

\item{...}{Additional arguments to barplot.}
}
\description{
Can plot positive and negative coefficients in two separate plots or the
same. X-axes are the same if separate plots are used.
}
