% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRisk.R
\name{plotRisk}
\alias{plotRisk}
\title{Plots the absolute w-year risk of individuals for different LM points for an event of interest within a window}
\usage{
plotRisk(
  superfm,
  data,
  format,
  LM_col,
  id_col,
  cause,
  varying,
  end_time,
  extend = F,
  silence = F,
  pch,
  lty,
  lwd,
  col,
  main,
  xlab,
  ylab,
  xlim,
  ...
)
}
\arguments{
\item{superfm}{Fitted LM super model}

\item{data}{Data frame of individuals from which to plot risk}

\item{format}{Character string specifying whether the data are in wide (default) or in long format}

\item{LM_col}{Character string specifying the column name in data containing the (running) time variable
associated with the time-varying covariate(s); only needed if format="long"}

\item{id_col}{Character string specifying the column name in data containing the subject id; only needed if format="long"}

\item{cause}{The cause we are looking at if considering competing risks}

\item{varying}{Character string specifying column name in the data containing time-varying covariates; only needed if format="wide"}

\item{end_time}{Final time point to plot risk}

\item{extend}{Argument to allow for risk to be plot at landmark times that are later than the LMs used in model fitting.
Default is FALSE. If set to TRUE, risks may be unreliable.}

\item{silence}{Silence the message when end_time > LMs used in fitting the model}

\item{pch}{Passed to points}

\item{lty}{Vector with line style}

\item{lwd}{Vector with line widths}

\item{col}{Vector with colors}

\item{main}{Title for the plot}

\item{xlab}{Label for x-axis}

\item{ylab}{Label for y-axis}

\item{xlim}{Limits for the x-axis}

\item{...}{Additional arguments passed to plot}
}
\value{
Single plot the absolute w-year risk of individuals
}
\description{
Plots the absolute w-year risk of individuals for different LM points for an event of interest within a window
}
