% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predLMrisk.R
\name{predLMrisk}
\alias{predLMrisk}
\title{Calculate w-year risk from a landmark time point}
\usage{
predLMrisk(
  object,
  newdata,
  tLM,
  cause,
  w,
  extend = F,
  silence = F,
  complete = T
)
}
\arguments{
\item{object}{fitted landmarking supermodel}

\item{newdata}{Either a dataframe of individuals to make predictions for which must contain the original covariates (i.e., without landmark interaction) or an object of class LMdataframe (e.g., from cutLMsuper and addLMtime).}

\item{tLM}{time points at which to predict risk of w more years.
Note tLM must be one value for newdata or must have the same length as the number of rows of newdata
(i.e., each datapoint is associated with one LM/prediction time point).
Alternatively, it can be a character string indicating a column in newdata.
It is only required when newdata is a dataframe.}

\item{cause}{Cause of interest if under competing risks.}

\item{w}{Prediction window, i.e., predict w-year (/month/..) risk from each of the tLMs.
Defaults to the w used in model fitting.
If w > than that used in model fitting, results are unreliable, but can be produced by setting extend=T.}

\item{extend}{Argument to allow for predictions at landmark times that are greater than those used in model fitting,
or prediction windows greater than the one used in model fitting.
Default is FALSE. If set to TRUE, predictions may be unreliable.}

\item{silence}{Silence the warning message when extend is set to TRUE.}

\item{complete}{Only make predictions for data entries with non-NA entries (i.e., non-NA predictions). Default is TRUE.}
}
\value{
An object of class "LMpred" with components:
\itemize{
\item preds: a dataframe with columns LM and risk, each entry corresponds to one individual and prediction time point (landmark)
\item w, type, LHS: as in the fitted super model
\item data: the newdata given in input
}
}
\description{
Calculate w-year risk from a landmark time point
}
\details{
See the Github for example code
}
\references{
van Houwelingen HC, Putter H (2012). Dynamic Prediction in Clinical Survival
Analysis. Chapman & Hall.
}
