% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.cv.pen_lm}
\alias{print.cv.pen_lm}
\title{Print the output from calling cv.pen_lm()\verb{, Similar to printing the output of printing a }cv.glmnet` object, print a
cross-validated penalized cause-specific Cox supermodel}
\usage{
\method{print}{cv.pen_lm}(x, all_causes = FALSE, silent = FALSE, digits = 3, ...)
}
\arguments{
\item{x}{a cv.pen_lm object}

\item{all_causes}{if \code{cv.pen_lm} fit a cause-specific Cox model, set TRUE to
print a summary of the glmnet path for each model.}

\item{silent}{Set TRUE to hide messages.}

\item{digits}{Number of significant digits to include}

\item{\dots}{additional print arguments}
}
\description{
Print the output from calling cv.pen_lm()\verb{, Similar to printing the output of printing a }cv.glmnet` object, print a
cross-validated penalized cause-specific Cox supermodel
}
\details{
If the model is a survival model (i.e., no competing risks), then the output
is the same as a call to glmnet would produce. For competing risks, the
default is only to print the output for the cause of interest (first cause).
Further events can be examined by setting \code{all_causes = TRUE}.
}
\references{
Friedman, J., Hastie, T. and Tibshirani, R. (2008).
Regularization Paths for Generalized Linear Models via Coordinate Descent
}
