% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.cv.penLM}
\alias{print.cv.penLM}
\title{Print the output from calling \code{cv.LMpen}, similar to the output of printing
a \code{cv.glmnet} object}
\usage{
\method{print}{cv.penLM}(x, all_causes = FALSE, silent = FALSE, digits = 3, ...)
}
\arguments{
\item{x}{a penLM object}

\item{all_causes}{if penLM fit a cause-specific Cox model, set TRUE to print
a summary of the glmnet path for each model.}

\item{silent}{Set TRUE to hide messages.}

\item{digits}{Number of significant digits to include}

\item{\dots}{additional print arguments}
}
\description{
I.e., print a cross-validated LMpen object
}
\details{
If the model is a survival model (i.e., no competing risks), then the output
is the same as a call to glmnet would produce. For competing risks, the
default is only to print the output for the cause of interest (first cause).
Further events can be examined by setting \code{all_causes = TRUE}.
}
\references{
Friedman, J., Hastie, T. and Tibshirani, R. (2008).
Regularization Paths for Generalized Linear Models via Coordinate Descent
}
