% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.LM.R
\name{print.LM.data.frame}
\alias{print.LM.data.frame}
\title{Print function for object of class LM.data.frame}
\usage{
\method{print}{LM.data.frame}(LMdata)
}
\arguments{
\item{LMdata}{Object of class LM.data.frame}
}
\value{
Printed output.
}
\description{
Print function for object of class LM.data.frame
}
\examples{

}
