% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.LMcoxph}
\alias{print.LMcoxph}
\title{Print function for object of class LMcoxph}
\usage{
\method{print}{LMcoxph}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{Object of class LMcoxph}

\item{verbose}{Boolean, default is FALSE. Print further components.}

\item{...}{Arguments passed to print.}
}
\value{
Printed output.
}
\description{
Print function for object of class LMcoxph
}
