% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.LMCSC}
\alias{print.LMCSC}
\title{Print function for object of class LMCSC}
\usage{
\method{print}{LMCSC}(x, verbose = FALSE, cause, ...)
}
\arguments{
\item{x}{Object of class LMCSC}

\item{verbose}{Boolean, default is FALSE. Print further components.}

\item{cause}{Print the model for a given cause.
If left out, all models are printed.}

\item{...}{Arguments passed to print.}
}
\value{
Printed output.
}
\description{
Print function for object of class LMCSC
}
