% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.LMpred}
\alias{print.LMpred}
\title{Print function for object of class LM.data.frame}
\usage{
\method{print}{LMpred}(x, ...)
}
\arguments{
\item{x}{Object of class LMpred}

\item{...}{Arguments passed to print.}
}
\value{
Printed output.
}
\description{
Print function for object of class LM.data.frame
}
