% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.penLMcoxph}
\alias{print.penLMcoxph}
\title{Print function for object of class penLMcoxph}
\usage{
\method{print}{penLMcoxph}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{Object of class penLMcoxph}

\item{verbose}{Logical, if verbose print func_covars, func_LMs, w, end_time
and type.}

\item{...}{Arguments passed to print.}
}
\value{
Printed output.
}
\description{
Print function for object of class penLMcoxph
}
