% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_lm_data.R
\name{sim_lm_data}
\alias{sim_lm_data}
\title{Title}
\usage{
sim_lm_data(
  supermodel,
  data,
  n,
  hazard.fun = NULL,
  replace = TRUE,
  censor = FALSE,
  multiplier = 1,
  tmax,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
