% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.cv.pen_lm}
\alias{summary.cv.pen_lm}
\title{Summarize cv.pen_lm objects: not yet implemented}
\usage{
\method{summary}{cv.pen_lm}(object, ...)
}
\arguments{
\item{object}{cv.pen_lm object}

\item{...}{Additional arguments}
}
\description{
Summarize cv.pen_lm objects: not yet implemented
}
