% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.dynamicLM}
\alias{summary.dynamicLM}
\title{Summarize dynamic_lm objects: not yet implemented}
\usage{
\method{summary}{dynamicLM}(object, ...)
}
\arguments{
\item{object}{dynamicLM object}

\item{...}{Additional arguments}
}
\description{
Summarize dynamic_lm objects: not yet implemented
}
