% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.pen_lm}
\alias{summary.pen_lm}
\title{Summarize pen_lm objects: not yet implemented}
\usage{
\method{summary}{pen_lm}(object, ...)
}
\arguments{
\item{object}{pen_lm object}

\item{...}{Additional arguments}
}
\description{
Summarize pen_lm objects: not yet implemented
}
