% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_metric.R
\name{summary_metric}
\alias{summary_metric}
\title{Obtain the summary metric from landmark-specific estimates}
\usage{
summary_metric(metric, df_t, df_c, df_iid, conf_int, object, id_col, B, se.fit)
}
\arguments{
\item{metric}{"AUC" or "Brier"}

\item{df_t}{time-dependent table of scores (AUC(s,t) or Brier(s,t)), i.e.,
a smaller table with entries for each landmark and model used for
landmark-estimates and their standard error

for example:
tLM      model   times      Brier          se      lower      upper b
1:   0 Null model 59.9999 0.06302245 0.008262163 0.04682891 0.07921599 1
2:   0  dynamicLM 59.9999 0.06140585 0.007931126 0.04586112 0.07695057 1
3:   6 Null model 65.9999 0.08492919 0.011250007 0.06287958 0.10697880 1
4:   6  dynamicLM 65.9999 0.08126512 0.010849048 0.06000138 0.10252887 1
5:  12 Null model 71.9999 0.11151696 0.013936209 0.08420249 0.13883143 1
6:  12  dynamicLM 71.9999 0.10554550 0.013541907 0.07900385 0.13208715 1}

\item{df_c}{table of contrast of scores.

for example:
tLM   times     model  reference  delta.Brier         se        lower        upper         p
1:   0 59.9999 dynamicLM Null model -0.001616602 0.00119091 -0.003950743 0.0007175397 0.1746381}

\item{df_iid}{df_iid contains the iid decomposition of the score, it is a
larger table with entries for each individual, who appear multiple times for
different landmarks and different models

for example:
tLM ID model   times      IF.Brier b
1:   0  1     0 59.9999 -5.853094e-02 1
2:   0  2     0 59.9999 -8.262274e-05 1
3:   0  3     0 59.9999 -5.852535e-02 1
...}

\item{conf_int}{Coverage level of the confidence interval.}

\item{object}{Either fitted supermodel or risk predictions.}

\item{id_col}{Column name of the ID (only needed if we incorporate
weighted metrics)}

\item{B}{Number of bootstrap replicates}

\item{se.fit}{If FALSE or 0, no standard errors are calculated.}
}
\value{
An list with entries \code{score} and optionally \code{contrasts} if
contrasts were calculated.
}
\description{
Obtain the summary metric from landmark-specific estimates
}
