% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_lm.R
\name{dynamic_lm}
\alias{dynamic_lm}
\title{Fit a dynamic Cox or cause-specific Cox landmark supermodel with or without
regularization.}
\usage{
dynamic_lm(...)
}
\arguments{
\item{...}{Arguments to pass to \code{dynamic_lm()}}
}
\value{
A fitted landmark supermodel object which has components:
\itemize{
\item model: fitted model
\item type: as input
\item w, func_covars, func_lms, lm_covs, all_covs, outcome: as in \code{lmdata}
\item LHS: the survival outcome
\item linear.predictors: the vector of linear predictors, one per subject.
Note that this vector has not been centered.
}

If the model is unpenalized (class "LMcoxph" or "LMCSC") it has additional
components:
\itemize{
\item args: arguments used to call model fitting
\item id_col: the cluster argument, often specifies the column with patient ID
\item lm_col: column name that indicates the landmark time point for a row.
}

If the model is penalized (class "penLMcoxph" or "penLMCSC") it has
additional components:
\itemize{
\item lambda: the values of lambda for which this model has been fit.
}
}
\description{
To fit Cox or cause-specific Cox models without regularization see:
\itemize{
\item \code{\link[=dynamic_lm.LMdataframe]{dynamic_lm.LMdataframe()}} for use on a stacked landmark dataset
\item \code{\link[=dynamic_lm.data.frame]{dynamic_lm.data.frame()}} for use on a dataframe
}

To fit penalized Cox or cause-specific Cox models see:
\itemize{
\item \code{\link[=dynamic_lm.pen_lm]{dynamic_lm.pen_lm()}} without cross-validation
\item \code{\link[=dynamic_lm.cv.pen_lm]{dynamic_lm.cv.pen_lm()}} with cross-validation
}
}
\seealso{
\code{\link[=dynamic_lm.LMdataframe]{dynamic_lm.LMdataframe()}}, \code{\link[=dynamic_lm.data.frame]{dynamic_lm.data.frame()}},
\code{\link[=dynamic_lm.pen_lm]{dynamic_lm.pen_lm()}}, \code{\link[=dynamic_lm.cv.pen_lm]{dynamic_lm.cv.pen_lm()}}
}
