% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_lm.R
\name{dynamic_lm}
\alias{dynamic_lm}
\title{Fit a dynamic Cox or cause-specific Cox landmark supermodel with or without
regularization.}
\usage{
dynamic_lm(...)
}
\value{
An object of class "LMcoxph", "LMCSC", "penLMcoxph" or "penLMCSC"
with components:
\itemize{
\item model: fitted model
\item type: as input
\item w, func_covars, func_lms, lm_covs, all_covs, outcome: as in \code{lmdata}
\item LHS: the survival outcome
\item linear.predictors: the vector of linear predictors, one per subject. Note
that this vector has not been centered.
\item args: arguments used to call model fitting                                 ####
\item id_col: the cluster argument, often specifies the column with patient ID   ####
\item lm_col: column name that indicates the landmark time point for a row       ####
}
}
\description{
To fit Cox or cause-specific Cox models without regularization see:
\itemize{
\item \code{\link[=dynamic_lm.LMdataframe]{dynamic_lm.LMdataframe()}} for use on a stacked landmark dataset
\item \code{\link[=dynamic_lm.data.frame]{dynamic_lm.data.frame()}} for use on a dataframe
}
}
\details{
To fit penalized Cox or cause-specific Cox models see:
\itemize{
\item \code{\link[=dynamic_lm.penLM]{dynamic_lm.penLM()}} without cross-validation
\item \code{\link[=dynamic_lm.cv.penLM]{dynamic_lm.cv.penLM()}} with cross-validation
}
}
\seealso{
\code{\link[=dynamic_lm.LMdataframe]{dynamic_lm.LMdataframe()}}, \code{\link[=dynamic_lm.data.frame]{dynamic_lm.data.frame()}},
\code{\link[=dynamic_lm.penLM]{dynamic_lm.penLM()}}, \code{\link[=dynamic_lm.cv.penLM]{dynamic_lm.cv.penLM()}}
}
