% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_lm.R
\name{dynamic_lm}
\alias{dynamic_lm}
\title{Fit a coxph or CSC model to a landmark super dataset, i.e., fit a dynamic
landmark supermodel}
\usage{
dynamic_lm(
  lmdata,
  formula,
  type = "coxph",
  method = "breslow",
  func_covars,
  func_lms,
  lm_col,
  outcome,
  w,
  lm_covs,
  cluster,
  x = FALSE,
  ...
)
}
\arguments{
\item{lmdata}{An object of class "LMdataframe", this can be created by
running \code{\link[=stack_data]{stack_data()}} and \code{\link[=add_interactions]{add_interactions()}}}

\item{formula}{The formula to be used, remember to include "+cluster(ID)" for
the column that indicates the ID of the individual for robust error
estimates.}

\item{type}{"coxph" or "CSC"/"CauseSpecificCox"}

\item{method}{A character string specifying the method for tie handling.
Default is "breslow". More information can be found in coxph.}

\item{func_covars}{A list of functions to use for interactions between LMs
and covariates.}

\item{func_lms}{A list of functions to use for transformations of the
landmark times.}

\item{lm_col}{Character string specifying the column name that indicates the
landmark time point for a row.}

\item{outcome}{List with items time and status, containing character strings
identifying the names of time and status variables, respectively, of the
survival outcome}

\item{w}{Scalar, the value of the prediction window (ie predict w-year/other
time period risk from the LM points)}

\item{lm_covs}{Vector of strings indicating the columns that are to have a
LM interaction}

\item{cluster}{Variable which clusters the observations (for e.g., identifies
repeated patient IDs), for the purposes of a robust variance.}

\item{x}{Logical value. If set to true, the \code{lmdata} is stored in the returned
object. This is required for internal validation.}

\item{...}{Arguments given to coxph or CSC.}
}
\value{
An object of class "LMcoxph" or "LMCSC" with components:
\itemize{
\item model: fitted model
\item type: as input
\item w, func_covars, func_lms, lm_covs, all_covs, outcome: as in \code{lmdata}
\item LHS: the LHS of the input formula
\item linear.predictors: the vector of linear predictors, one per subject.
Note that this vector has not been centered.
\item args: arguments used to call model fitting
\item id_col: the cluster argument, often specifies the column with patient ID
\item lm_col: column name that indicates the landmark time point for a row.
}
}
\description{
dynamic (dyn) landmark (lm) supermodel --> dynamic_lm
}
\examples{
\dontrun{
data(relapse)
outcome <- list(time = "Time", status = "event")
covars <- list(fixed = c("ID","age.at.time.0","male","stage","bmi"),
               varying = c("treatment"))
w <- 60; lms <- c(0, 6, 12, 18)
LMs = seq(0,36,by=6)
# Covariate-landmark time interactions
func_covars <- list(function(t) t, function(t) t^2)
# let hazard depend on landmark time
func_lms <- list(function(t) t, function(t) t^2)
# Choose covariates that will have time interaction
pred_covars <- c("age","male","stage","bmi","treatment")
# Stack landmark datasets
lmdata <- stack_data(relapse, outcome, lms, w, covars, format = "long",
                     id = "ID", rtime = "T_txgiven")
# Update complex LM-varying covariates, note age is in years and LM is in months
lmdata$data$age <- lmdata$data$age.at.time.0 + lmdata$data$LM/12
# Add LM-time interactions
lmdata <- add_interactions(lmdata, pred_covars, func_covars, func_lms)
head(lmdata$data)

formula <- "Hist(Time, event, LM) ~ age + male + stage + bmi + treatment +
           age_1 + age_2 + male_1 + male_2 + stage_1 + stage_2 + bmi_1 +
           bmi_2 + treatment_1 + treatment_2 + LM_1 + LM_2 + cluster(ID)"
supermodel <- dynamic_lm(lmdata, as.formula(formula), "CSC")
print(supermodel)

par(mfrow = c(2,3))
plot(supermodel)
}
}
