% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLM.R
\name{fitLM}
\alias{fitLM}
\title{Fit a coxph or CSC model to a LM super dataset}
\usage{
fitLM(
  formula,
  lmdata,
  type = "coxph",
  method = "breslow",
  func_covars,
  func_lms,
  LM_col,
  outcome,
  w,
  LMcovars,
  cluster,
  x = FALSE,
  ...
)
}
\arguments{
\item{formula}{The formula to be used, remember to include "+cluster(ID)" for
the column that indicates the ID of the individual for robust error
estimates.}

\item{lmdata}{An object of class "LMdataframe", this can be created by
running \code{\link[=stack_data]{stack_data()}} and \code{\link[=add_interactions]{add_interactions()}}}

\item{type}{"coxph" or "CSC"/"CauseSpecificCox"}

\item{method}{A character string specifying the method for tie handling.
Default is "breslow". More information can be found in coxph.}

\item{func_covars}{A list of functions to use for interactions between LMs
and covariates.}

\item{func_lms}{A list of functions to use for transformations of the
landmark times.}

\item{LM_col}{Character string specifying the column name that indicates the
landmark time point for a row.}

\item{outcome}{List with items time and status, containing character strings
identifying the names of time and status variables, respectively, of the
survival outcome}

\item{w}{Scalar, the value of the prediction window (ie predict w-year/other
time period risk from the LM points)}

\item{LMcovars}{Vector of strings indicating the columns that are to have a
LM interaction}

\item{cluster}{Variable which clusters the observations (for e.g., identifies
repeated patient IDs), for the purposes of a robust variance.}

\item{x}{Logical value. If set to true, the \code{lmdata} is stored in the returned
object. This is required for internal validation.}

\item{...}{Arguments given to coxph or CSC.}
}
\value{
An object of class "LMcoxph" or "LMCSC" with components:
\itemize{
\item model: fitted model
\item type: as input
\item w, func_covars, func_lms, LMcovars, allLMcovars, outcome: as in \code{lmdata}
\item LHS: the LHS of the input formula
\item linear.predictors: the vector of linear predictors, one per subject.
Note that this vector has not been centered.
\item args: arguments used to call model fitting
\item ID_col: the cluster argument, often specifies the column with patient ID
}
}
\description{
Fit a coxph or CSC model to a LM super dataset
}
\examples{
\dontrun{
data(relapse)
outcome = list(time="Time", status="event")
covars = list(fixed=c("ID","age.at.time.0","male","stage","bmi"),
              varying=c("treatment"))
w = 60; LMs = c(0,12,24)
# Covariate-landmark time interactions
func.covars <- list( function(t) t, function(t) t^2)
# let hazard depend on landmark time
func.LMs <- list( function(t) t, function(t) t^2)
# Choose covariates that will have time interaction
pred.covars <- c("age","male","stage","bmi","treatment")
# Stack landmark datasets
lmdata <- stack_data(relapse, outcome, LMs, w, covs, format="long",
                     id="ID", rtime="fup_time", right=F)
# Update complex LM-varying covariates, note age is in years and LM is in months
lmdata$data$age <- lmdata$data$age.at.time.0 + lmdata$data$LM/12
# Add LM-time interactions
lmdata <- add_interactions(lmdata, pred.covars, func.covars, func.LMs)
formula <- "Hist(Time, event, LM) ~ age + male + stage + bmi + treatment +
           age_1 + age_2 + male_1 + male_2 + stage_1 + stage_2 + bmi_1 +
           bmi_2 + treatment_1 + treatment_2 + LM_1 + LM_2 + cluster(ID)"
supermodel <- fitLM(as.formula(formula), lmdata, "CSC")
}
}
