% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lm_data.R
\name{get_lm_data}
\alias{get_lm_data}
\title{Build a landmark dataset}
\usage{
get_lm_data(
  data,
  outcome,
  lm,
  horizon,
  covs,
  format = c("wide", "long"),
  id,
  rtime,
  right = TRUE
)
}
\arguments{
\item{data}{Data frame from which to construct landmark super dataset}

\item{outcome}{A list with items time and status, containing character
strings identifying the names of time and status variables, respectively, of the survival outcome}

\item{lm}{The value of the landmark time point at which to construct the
landmark dataset.}

\item{horizon}{Scalar, the value of the prediction window (ie predict risk
within time w landmark points)}

\item{covs}{A list with items fixed and varying, containing character strings
specifying column names in the data containing time-fixed and time-varying
covariates, respectively.}

\item{format}{Character string specifying whether the original data are in
wide (default) or in long format.}

\item{id}{Character string specifying the column name in data containing the
subject id.}

\item{rtime}{Character string specifying the column name in data containing
the (running) time variable associated with the time-varying variables;
only needed if format = "long".}

\item{right}{Boolean (default = FALSE), indicating if the intervals for the
time-varying covariates are closed on the right (and open on the left) or
vice-versa.}
}
\value{
A landmark dataset.
}
\description{
Build a landmark dataset
}
\details{
This function is based from \code{\link[dynpred:cutLM]{dynpred::cutLM()}} with minor changes.
The original function was authored by Hein Putter.
}
\references{
van Houwelingen HC, Putter H (2012). Dynamic Prediction in
Clinical Survival Analysis. Chapman & Hall.
}
