% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMcalPlot.R
\name{LMcalPlot}
\alias{LMcalPlot}
\title{Calibration plots for dynamic risk prediction landmark models}
\usage{
LMcalPlot(
  preds,
  unit = "year",
  cause,
  tLM,
  formula,
  plot = T,
  main,
  sub = T,
  ...
)
}
\arguments{
\item{preds}{A named list of prediction models, where allowed entries are outputs from predLMrisk}

\item{unit}{The unit of w, i.e. w-unit prediction ("year","month", etc...). Used to label the plot.}

\item{cause}{Cause of interest if considering competing risks}

\item{tLM}{Landmark times for which calibration must be plot. These must be a subset of LM times used during the prediction}

\item{formula}{A survival or event history formula. The left hand side is used to compute the expected event status.
If none is given, it is obtained from the prediction object.}

\item{plot}{If FALSE, do not plot the results, just return a plottable object.}

\item{main}{Optional title to override default.}

\item{sub}{If TRUE, add a subheading with the number of individuals at risk, and the number that under the event of interest}

\item{...}{Additional arguments to pass to calPlot}
}
\value{
List of plots of w-year risk, one entry per prediction/landmark time point
}
\description{
Calibration plots for dynamic risk prediction landmark models
}
\details{
See the Github for example code
}
