% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMcalPlot.R
\name{LMcalPlot}
\alias{LMcalPlot}
\title{Calibration plots for dynamic risk prediction landmark models}
\usage{
LMcalPlot(
  preds,
  unit = "year",
  cause,
  tLM,
  formula,
  plot = T,
  main,
  sub = T,
  splitMethod = "none",
  B = 1,
  ...
)
}
\arguments{
\item{preds}{A named list of prediction models, where allowed entries are outputs from predLMrisk}

\item{unit}{The unit of w, i.e. w-unit prediction ("year","month", etc...). Used to label the plot.}

\item{cause}{Cause of interest if considering competing risks}

\item{tLM}{Landmark times for which calibration must be plot. These must be a subset of LM times used during the prediction}

\item{formula}{A survival or event history formula. The left hand side is used to compute the expected event status.
It is recommended to give a formula. If none is given, it is obtained from the prediction object.}

\item{plot}{If FALSE, do not plot the results, just return a plottable object. Default is TRUE.}

\item{main}{Optional title to override default.}

\item{sub}{If TRUE, add a subheading with the number of individuals at risk, and the number that under the event of interest.
Default is TRUE.}

\item{splitMethod}{Defines the internal validation design as in pec::calPlot. Options are none/noPlan or BootCv.}

\item{B}{The number of cross-validation steps.}

\item{...}{Additional arguments to pass to calPlot}
}
\value{
List of plots of w-year risk, one entry per prediction/landmark time point
}
\description{
Calibration plots for dynamic risk prediction landmark models
}
\details{
Most errors in plotting occur when a formula is not given. Formulas can look like \code{Surv(LM,Time,event)~1} / \code{Surv(LM,Time,event==1)~1} / \code{Hist(Time,event,LM)~1} / similar...

See the Github for example code on using LMcalPlot in general.
}
