% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMScore.R
\name{LMScore}
\alias{LMScore}
\title{Methods (AUCt, Brier) to score the predictive performance of dynamic risk markers from LM super models}
\usage{
LMScore(
  preds,
  formula,
  metrics = c("auc", "brier"),
  cause,
  tLM,
  unit,
  split.method,
  B,
  ...
)
}
\arguments{
\item{preds}{A named list of prediction models, where allowed entries are outputs from predLMrisk}

\item{formula}{A survival or event history formula. The left hand side is used to compute the expected event status.
If none is given, it is obtained from the prediction object.}

\item{metrics}{Character vector specifying which metrics to apply. Choices are "auc" and "brier". Case matters.}

\item{cause}{Cause of interest if considering competing risks}

\item{tLM}{Landmark times for which scores must be given. These must be a subset of LM times used during the prediction}

\item{unit}{Time unit for window of prediction, e.g., "year", "month", etc. Used for printing results.}

\item{split.method}{Method for cross-validation. Right now, as in riskRegression, the only option is bootcv.}

\item{B}{The number of bootstap steps for cross-validation.}

\item{...}{Additional arguments to pass to Score (riskRegression package)}
}
\value{
An object of class "LMScore", which has components:
\itemize{
\item auct: dataframe containing time-dependent auc information if "auc" was a metric
\item briert: dataframe containing time-dependent brier score if "brier" was a metric
}
}
\description{
Methods (AUCt, Brier) to score the predictive performance of dynamic risk markers from LM super models
}
\details{
See the Github for example code
}
