% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pen_lm.R
\name{pen_lm}
\alias{pen_lm}
\title{Compute the regularization path of coefficients for a Cox or
cause-specific Cox landmark supermodel with lasso or elasticnet penalization.}
\usage{
pen_lm(x, y, alpha = 1, ...)
}
\arguments{
\item{x}{An "LMdataframe", which can be created by running \code{\link[=stack_data]{stack_data()}} and
\code{\link[=add_interactions]{add_interactions()}}.}

\item{y}{Optional, a vector of column names of the data stored in \code{lmdata}
that are to be used as dependent variables. If not specified, it is assumed
that all non-response variables are the dependent variables.}

\item{alpha}{The elastic net mixing parameter: Lies between 0 and 1. At 1,
the penalty is the LASSO penalty, and at 0, the penalty is the ridge
penalty. The default is 1.}

\item{...}{Additional arguments passed to \code{glmnet()}.}
}
\value{
An object of class \code{pen_lm}. This is a list of \code{glmnet} objects (one
for each cause-specific Cox model or a list of length one for a regular Cox
model). The object also has attributes \code{survival.type} (\code{competing.risk}
or \code{survival}) and \code{lmdata} and \code{xcols} which store the inputs if given.
Functions \code{print} and \code{plot} exist for the object. To make predictions,
see \code{\link[=dynamic_lm.pen_lm]{dynamic_lm.pen_lm()}} and \code{\link[=predict.dynamicLM]{predict.dynamicLM()}}.
}
\description{
Fit by calling \verb{[glmnet::glmnet()]}. As in \code{glmnet}, the model is fit via penalized
maximum likelihood to produce a regularization path at a grid of values for
the regularization parameter lambda. Input can be as typically done for
\code{glmnet} in the form of \code{x} and \code{y} which are a matrix and response object
or with a landmark super dataset specifying the dependent columns in \code{y}.
}
\examples{
\dontrun{
}
}
\seealso{
\code{\link[=print.pen_lm]{print.pen_lm()}}, \code{\link[=plot.pen_lm]{plot.pen_lm()}}, \code{\link[=dynamic_lm.pen_lm]{dynamic_lm.pen_lm()}}
}
