% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.coefs}
\alias{plot.coefs}
\title{Generic function to plot coefficients}
\usage{
\method{plot}{coefs}(x, single.plot, max_coefs, ...)
}
\arguments{
\item{x}{(Named) Vector of coefficients}

\item{single.plot}{Logical, defaults to FALSE. A single plot for both
positive and negative coefficients, or two separate plots.}

\item{max_coefs}{Default is 10. The maximum number of coefficients to plot.
Can be set to NULL if plotting all coefficients is desired.}

\item{...}{Additional arguments to barplot.}
}
\description{
Can plot positive and negative coefficients in two separate plots or the
same. X-axes are the same if separate plots are used.
}
