% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.dynamicLM}
\alias{plot.dynamicLM}
\title{Plots the dynamic log-hazard ratio of a cox or CSC supermodel}
\usage{
\method{plot}{dynamicLM}(
  object,
  covars,
  conf_int = TRUE,
  cause,
  end_time,
  logHR = TRUE,
  extend = FALSE,
  silence = FALSE,
  xlab = "LM time",
  ylab,
  ylim,
  main,
  ...
)
}
\arguments{
\item{object}{An object of class "LMcoxph" or "LMCSC", i.e. a fitted
supermodel}

\item{covars}{Vector or list of strings indicating the variables to plot
(note these must be given without time interaction label, for e.g., as in
the argument \code{lm_covs} in \code{\link[=add_interactions]{add_interactions()}}).}

\item{conf_int}{Include confidence intervals or not, default is TRUE}

\item{cause}{Cause of interest if considering competing risks}

\item{end_time}{Final time point to plot HR, defaults to the last landmark
point used in model fitting.}

\item{logHR}{Boolean, if true plots the log of the hazard ratio, if false
plots the hazard ratio. Default is TRUE.}

\item{extend}{Argument to allow for HR to be plot at landmark times that are
later than the LMs used in model fitting.
Default is FALSE. If set to TRUE, the HR may be unreliable.}

\item{silence}{silence the warning message when end_time > LMs used in
fitting the model}

\item{xlab}{x label for the plots}

\item{ylab}{y label for the plots}

\item{ylim}{y limit for the plots}

\item{main}{Vector of strings indicating the title of each plot. Must be in
the same order as covars.}

\item{...}{Additional arguments passed to plot}
}
\value{
Plots for each variable in covars showing the dynamic hazard ratio
}
\description{
Plots the dynamic log-hazard ratio of a cox or CSC supermodel
}
\details{
See our \href{https://github.com/thehanlab/dynamicLM}{GitHub} for example
code
}
