% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.LMScore}
\alias{plot.LMScore}
\title{Plot an object output from \code{\link[=score]{score()}}: plot the landmark and time-dependent
Brier and/or AUC of dynamic landmark supermodels.}
\usage{
\method{plot}{LMScore}(x, metrics, xlab, ylab, loc, pch, ...)

\method{plot}{LMScore}(x, metrics, xlab, ylab, loc, pch, ...)
}
\arguments{
\item{x}{An object of class "LMScore" output from \code{score}}

\item{metrics}{One or both of "auc" and "brier"}

\item{xlab, ylab, pch, ylim, xlim}{graphical parameters}

\item{ylab}{y label}

\item{loc}{legend location}

\item{pch}{size of points}

\item{...}{}

\item{se}{Boolean, default TRUE. To include point wise confidence intervals.}

\item{xlab}{x label}
}
\description{
Plot an object output from \code{\link[=score]{score()}}: plot the landmark and time-dependent
Brier and/or AUC of dynamic landmark supermodels.

Plot an object output from \code{\link[=score]{score()}}: plot the landmark and time-dependent
Brier and/or AUC of dynamic landmark supermodels.
}
