% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.LMScore}
\alias{plot.LMScore}
\title{Plot an object output from \code{LMScore}: plot the landmark and time-dependent
Brier and/or AUC of dynamic landmark supermodels.}
\usage{
\method{plot}{LMScore}(object, metrics, xlab, ylab, x, pch, ...)
}
\arguments{
\item{object}{An object of class "LMScore" output from \code{LMScore}}

\item{metrics}{One or both of "auc" and "brier"}

\item{xlab}{x label}

\item{ylab}{y label}

\item{x}{legend location}

\item{pch}{size of points}

\item{...}{}
}
\description{
Plot an object output from \code{LMScore}: plot the landmark and time-dependent
Brier and/or AUC of dynamic landmark supermodels.
}
