% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.pen_lm}
\alias{plot.pen_lm}
\title{Plot coefficients from an object created by calling \code{\link[=pen_lm]{pen_lm()}}, analogous to
plotting from \code{\link[=glmnet]{glmnet()}}}
\usage{
\method{plot}{pen_lm}(
  x,
  xvar = "norm",
  all_causes = FALSE,
  silent = FALSE,
  label = FALSE,
  all_causes_title = TRUE,
  ...
)
}
\arguments{
\item{x}{a fitted \code{\link[=pen_lm]{pen_lm()}} object}

\item{xvar}{As in \code{\link[=glmnet]{glmnet()}}: "What is on the X-axis. "\code{norm}" plots against
the L1-norm of the coefficients, "\code{lambda}" against the log-lambda
sequence, and "\code{dev}" against the percent deviance explained."}

\item{all_causes}{if pen_lm fit a cause-specific Cox model, set TRUE to plot
coefficient profile plots for each model.}

\item{silent}{Set TRUE to hide messages.}

\item{label}{Set TRUE to label the curves by variable index numbers.}

\item{all_causes_title}{If \code{all_causes} is set to TRUE, includes a title with
the cause. Defaults to TRUE.}

\item{\dots}{additional graphical parameters}
}
\description{
As in the glmnet package, produces a coefficient profile plot of the
coefficient paths.
}
\details{
If the model is a survival model (i.e., no competing risks), then the output
is the same as a call to \code{glmnet} would produce. For competing risks, the
default is only to plot the coefficient profile plot for the cause of
interest (first cause) Further events can be examined by setting
\code{all_causes = TRUE}.
}
