% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.penLMcoxph}
\alias{plot.penLMcoxph}
\title{Plot the non-zero coefficients of a penalized Cox landmark supermodel or
the dynamic log-hazard ratios}
\usage{
\method{plot}{penLMcoxph}(
  x,
  single_plot = TRUE,
  max_coefs = NULL,
  col = "blue",
  xlab = "Coefficient value",
  HR = FALSE,
  covars = NULL,
  ...
)
}
\arguments{
\item{x}{a penalized Cox supermodel - created by calling \code{\link[=dynamic_lm]{dynamic_lm()}} on an
object created from \code{\link[=pen_lm]{pen_lm()}} or \code{\link[=cv.pen_lm]{cv.pen_lm()}}.}

\item{single_plot}{Logical, defaults to TRUE. A single plot for both
positive and negative coefficients, or two separate plots.}

\item{max_coefs}{Default is to plot all coefficients. If specified, gives the
maximum number of coefficients to plot.}

\item{col}{Fill color for the barplot.}

\item{xlab}{x-axis Label}

\item{HR}{Plot the hazard ratio? Default is FALSE. See \code{\link[=plot.dynamicLM]{plot.dynamicLM()}}
for additional arguments.}

\item{covars}{If HR is TRUE, a vector or list of strings indicating the
variables to plot (note these must be given without time interaction).
Defaults to all non-zero variables.}

\item{...}{Additional arguments to barplot or to \code{\link[=plot.dynamicLM]{plot.dynamicLM()}}.}
}
\description{
Can plot positive and negative coefficients in two separate plots or the
same. X-axes are the same if separate plots are used.
}
\details{
If plotting the log hazard ratios, check \code{\link[=plot.dynamicLM]{plot.dynamicLM()}} to
see further arguments.
}
