% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.penLMCSC}
\alias{plot.penLMCSC}
\title{Plot the coefficients of a penalized Cause-specific Cox supermodel}
\usage{
\method{plot}{penLMCSC}(object, single.plot = FALSE, max_coefs = 10, all_causes = FALSE, ...)
}
\arguments{
\item{object}{a penalized cause-specific Cox supermodel - created by calling
\code{fitLM} on an object created from \code{penLM}/\code{cv.pen}}

\item{single.plot}{Logical, defaults to FALSE. A single plot for both
positive and negative coefficients, or two separate plots.}

\item{max_coefs}{Default is 10. The maximum number of coefficients to plot.
Can be set to NULL if plotting all coefficients is desired.}

\item{all_causes}{Logical, default is FALSE. Plot coefficients for all
cause-specific models.}

\item{...}{Additional arguments to barplot.}
}
\description{
Plot the coefficients of a penalized Cause-specific Cox supermodel
}
