% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotrisk.R
\name{plotrisk}
\alias{plotrisk}
\title{Plots the absolute risk of individuals for different LM points for an event
of interest within a given window}
\usage{
plotrisk(
  object,
  data,
  format,
  lm_col,
  id_col,
  w,
  cause,
  varying,
  end_time,
  extend = FALSE,
  silence = FALSE,
  pch,
  lty,
  lwd,
  col,
  main,
  xlab,
  ylab,
  xlim,
  ylim = c(0, 1),
  x.legend,
  y.legend,
  las = 1,
  ...
)
}
\arguments{
\item{object}{Fitted landmark supermodel}

\item{data}{Data frame of individuals from which to plot risk}

\item{format}{Character string specifying whether the data are in wide
(default) or in long format}

\item{lm_col}{Character string specifying the column name in data containing
the (running) time variable
associated with the time-varying covariate(s); only needed if format="long"}

\item{id_col}{Character string specifying the column name in data containing
the subject id; only needed if format="long"}

\item{w}{Prediction window, i.e., predict w-year (/month/..) risk from each
of the tLMs.
Defaults to the w used in model fitting.
If w > than that used in model fitting, results are unreliable, but can be
produced by setting extend=T.}

\item{cause}{The cause we are looking at if considering competing risks}

\item{varying}{Character string specifying column name in the data containing
time-varying covariates; only needed if format="wide"}

\item{end_time}{Final time point to plot risk}

\item{extend}{Argument to allow for risk to be plot at landmark times that
are later than the landmarks used in model fitting.
Default is FALSE. If set to TRUE, risks may be unreliable.}

\item{silence}{Silence the message when end_time > landmarks used in fitting
the model}

\item{pch}{Passed to points}

\item{lty}{Vector with line style}

\item{lwd}{Vector with line widths}

\item{col}{Vector with colors}

\item{main}{Title for the plot}

\item{xlab}{Label for x-axis}

\item{ylab}{Label for y-axis}

\item{xlim}{Limits for the x-axis}

\item{ylim}{Limits for the y-axis}

\item{x.legend, y.legend}{The x and y co-ordinates to be used to position the
legend. They can be specified by keyword or in any way which is accepted
by xy.coords.}

\item{las}{the style of axis labels}

\item{...}{Additional arguments passed to plot}
}
\value{
Single plot of the absolute w-year risk of individuals
}
\description{
Plots the absolute risk of individuals for different LM points for an event
of interest within a given window
}
\details{
See our \href{https://github.com/thehanlab/dynamicLM}{GitHub} for example
code
}
\examples{
\dontrun{
data(relapse)

# select patients whose risk we want to plot
idx <- relapse$ID \%in\% c("ID1007", "ID1301")

outcome <- list(time = "Time", status = "event")
covars <- list(fixed = c("age.at.time.0", "male", "stage", "bmi"),
               varying = c("treatment"))
w <- 60; lms <- c(0, 6, 12, 18)

# Prediction time points (how smooth the plot will be)
x <- seq(0, 18, by = 1)

# Stack landmark datasets
dat <- stack_data(relapse[idx, ], outcome, x, w, covars, format = "long",
                  id = "ID", rtime = "T_txgiven")$data
dat$age <- dat$age.at.time.0 + dat$LM / 12 # age is in years and LM is in months
head(dat)

plotrisk(supermodel, dat, format = "long", ylim = c(0, 1), #0.7),
         x.legend = "bottomright")
}

}
