% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.dynamicLM}
\alias{predict.dynamicLM}
\title{Calculate w-year risk from a landmark time point}
\usage{
\method{predict}{dynamicLM}(
  object,
  newdata,
  lms,
  cause,
  w,
  extend = FALSE,
  silence = FALSE,
  complete = TRUE,
  ...
)
}
\arguments{
\item{object}{Fitted landmark supermodel}

\item{newdata}{Either a dataframe of individuals to make predictions for or
an object of class LMdataframe (e.g., created by calling \code{\link[=stack_data]{stack_data()}} and
\code{\link[=add_interactions]{add_interactions()}}). If it is a dataframe, it must contain the
original covariates (i.e., without landmark interaction).}

\item{lms}{landmark time points that correspond to the entries in \code{newdata}.
Only required when \code{newdata} is a data.frame.
\code{lms} is either a time point, a vector or character string.
\itemize{
\item For a single time point, w-year risk is predicted from this time for each
data point.
\item For a vector, \code{lms} must have the same length as the number of rows of
\code{newdata} (i.e., each data point is associated with one LM/prediction
time point).
\item A character string indicates a column in \code{newdata}.
}}

\item{cause}{Cause of interest for competing risks.}

\item{w}{Prediction window, i.e., predict w-year (/month/..) risk from each
of the \code{lms}. Defaults to the w used in model fitting.
If \code{w} > than that used in model fitting, results are unreliable, but can
be produced by setting \code{extend = T}.}

\item{extend}{Argument to allow for predictions at landmark times that are
later than those used in model fitting, or prediction windows greater
than the one used in model fitting.
Default is FALSE. If set to TRUE, predictions may be unreliable.}

\item{silence}{Silence the warning message when extend is set to TRUE.}

\item{complete}{Only make predictions for data entries with non-NA entries
(i.e., non-NA predictions). Default is TRUE.}

\item{...}{Unimplemented for now.}
}
\value{
An object of class "LMpred" with components:
\itemize{
\item preds: a dataframe with columns LM and risk, each entry corresponds to
one individual and prediction time point (landmark)
\item w, type, LHS: as in the fitted super model
\item data: the newdata given in input
}
}
\description{
Calculate w-year risk from a landmark time point
}
\examples{
\dontrun{
data(relapse)
outcome <- list(time = "Time", status = "event")
covars <- list(fixed = c("age.at.time.0", "male", "stage", "bmi"),
               varying = c("treatment"))
w <- 60; lms <- c(0, 6, 12, 18)
# Choose covariates that will have time interaction
pred_covars <- c("age", "male", "stage", "bmi", "treatment")
# Stack landmark datasets
lmdata <- stack_data(relapse, outcome, lms, w, covars, format = "long",
                     id = "ID", rtime = "T_txgiven")

# Update complex landmark-varying covariates
# note age is in years and LM is in months
lmdata$data$age <- lmdata$data$age.at.time.0 + lmdata$data$LM/12
# Add LM-time interactions
lmdata <- add_interactions(lmdata, pred_covars,
                           func_covars = c("linear", "quadratic"),
                           func_lms = c("linear", "quadratic"))

formula <- "Hist(Time, event, LM) ~ age + male + stage + bmi + treatment +
           age_1 + age_2 + male_1 + male_2 + stage_1 + stage_2 + bmi_1 +
           bmi_2 + treatment_1 + treatment_2 + LM_1 + LM_2 + cluster(ID)"
supermodel <- dynamic_lm(lmdata, as.formula(formula), "CSC")

p1 <- predict(supermodel)
head(p1$preds)
}
}
\references{
van Houwelingen HC, Putter H (2012). Dynamic Prediction in
Clinical Survival Analysis. Chapman & Hall.
}
\seealso{
\code{\link[=stack_data]{stack_data()}}, \code{\link[=add_interactions]{add_interactions()}}, \code{\link[=dynamic_lm]{dynamic_lm()}}, \code{\link[=score]{score()}},
\code{\link[=calplot]{calplot()}}
}
