% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.LMCSC}
\alias{print.LMCSC}
\title{Print function for object of class LMCSC}
\usage{
\method{print}{LMCSC}(x, ...)
}
\arguments{
\item{x}{Object of class LMCSC}

\item{...}{Arguments passed to print.}
}
\value{
Printed output.
}
\description{
Print function for object of class LMCSC
}
