% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.LMdataframe}
\alias{print.LMdataframe}
\title{Print function for object of class LMdataframe}
\usage{
\method{print}{LMdataframe}(x, ...)
}
\arguments{
\item{x}{Object of class LMdataframe}

\item{...}{Arguments passed to print.}
}
\value{
Printed output.
}
\description{
Print function for object of class LMdataframe
}
