% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.LMpred}
\alias{print.LMpred}
\title{Print function for object of class LMpred}
\usage{
\method{print}{LMpred}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{Object of class LMpred}

\item{verbose}{Boolean, default is FALSE. Print further components.}

\item{...}{Arguments passed to print.}
}
\value{
Printed output.
}
\description{
Print function for object of class LMpred
}
