% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.LMScore}
\alias{print.LMScore}
\title{Print function for object of class LMScore}
\usage{
\method{print}{LMScore}(x, digits = 3, ...)
}
\arguments{
\item{x}{Object of class LMScore}

\item{digits}{Number of significant digits to include}

\item{...}{Arguments passed to print.}
}
\value{
Printed output.
}
\description{
Print function for object of class LMScore
}
