% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskScore.R
\name{riskScore}
\alias{riskScore}
\title{Calcutes dynamic risk score at a time for an individual (helper to
predict.dynamicLM)}
\usage{
riskScore(object, tLM, data, func_covars, func_lms)
}
\arguments{
\item{object}{A coxph object}

\item{tLM}{Landmarking time point at which to calculate risk score (time at which the prediction is made)}

\item{data}{Dataframe (single row) of individual. Must contain the original covariates.}

\item{func_covars}{A list of functions to use for interactions between LMs and covariates.}

\item{func_lms}{A list of functions to use for transformations of the landmark times.}
}
\value{
Numeric risk score
}
\description{
Calcutes dynamic risk score at a time for an individual (helper to
predict.dynamicLM)
}
