% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_data.R
\name{stack_data}
\alias{stack_data}
\title{Build a stacked super dataset from original dataset (wide or long format)}
\usage{
stack_data(
  data,
  outcome,
  lms,
  w,
  covs,
  format = c("wide", "long"),
  id,
  rtime,
  right = F
)
}
\arguments{
\item{data}{Data frame from which to construct landmark super dataset}

\item{outcome}{List with items time and status, containing character strings identifying the names of time and status variables, respectively, of the survival outcome}

\item{lms}{vector, the value of the landmark time points (points at which prediction is made)}

\item{w}{Scalar, the value of the prediction window (ie predict w-year/other time period risk from the LM points)}

\item{covs}{List with items fixed and varying, containing character strings specifying column names in the data containing time-fixed and time-varying covariates, respectively}

\item{format}{Character string specifying whether the original data are in wide (default) or in long format}

\item{id}{Character string specifying the column name in data containing the subject id; only needed if format="long"}

\item{rtime}{Character string specifying the column name in data containing the (running) time variable associated; only needed if format="long"}

\item{right}{Boolean (default=TRUE), indicating if the intervals for the time-varying covariates are closed on the right (and open on the left) or vice versa, see cut}
}
\value{
An object of class "LMdataframe". This the following components:
\itemize{
\item data: containing the stacked data set, i.e., the outcome and the values of time-fixed and time-varying covariates taken at the landmark time points. The value of the landmark time point is stored in column LM.
\item outcome: same as input
\item w: same as input
\item end_time: final landmarking point used in training
}
}
\description{
Build a stacked super dataset from original dataset (wide or long format)
}
\details{
This function calls cutLM from the library dynpred, more documentation can be found there. Note that for every landmark tLM given in \code{lms}, there must be at least one patient alive after tLM.
}
\examples{
\dontrun{
head(lmdata$data)
data(relapse)
outcome <- list(time = "Time", status = "event")
covars <- list(fixed = c("ID","age.at.time.0","male","stage","bmi"),
               varying = c("treatment"))
w <- 60; lms <- c(0, 6, 12, 18)
# Stack landmark datasets
lmdata <- stack_data(relapse, outcome, lms, w, covars, format = "long",
                     id = "ID", rtime = "T_txgiven")
head(lmdata$data)
}
}
